+++ lib/src/ast/selector.dart
@@ -3,6 +3,7 @@
 // https://opensource.org/licenses/MIT.
 
 import 'package:meta/meta.dart';
+import 'package:source_span/source_span.dart';
 
 import '../deprecation.dart';
 import '../evaluation_context.dart';
@@ -10,6 +11,7 @@ import '../exception.dart';
 import '../visitor/any_selector.dart';
 import '../visitor/interface/selector.dart';
 import '../visitor/serialize.dart';
+import 'node.dart';
 import 'selector/complex.dart';
 import 'selector/list.dart';
 import 'selector/placeholder.dart';
@@ -39,7 +41,7 @@ export 'selector/universal.dart';
 /// Selectors have structural equality semantics.
 ///
 /// {@category AST}
-abstract class Selector {
+abstract class Selector implements AstNode {
   /// Whether this selector, and complex selectors containing it, should not be
   /// emitted.
   ///
@@ -77,10 +79,14 @@ abstract class Selector {
   @internal
   bool get isUseless => accept(const _IsUselessVisitor());
 
+  final FileSpan span;
+
+  Selector(this.span);
+
   /// Prints a warning if [this] is a bogus selector.
   ///
   /// This may only be called from within a custom Sass function. This will
-  /// throw a [SassScriptException] in Dart Sass 2.0.0.
+  /// throw a [SassException] in Dart Sass 2.0.0.
   void assertNotBogus({String? name}) {
     if (!isBogus) return;
     warn(